+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 [^format]
 
-<!-- todo support file uploads -->
+<!-- Todo support file uploads -->